+++ client/src/game/systems/logic/door/events.ts
@@ -4,6 +4,8 @@ import { getLocalId } from "../../../id";
 import type { GlobalId } from "../../../id";
 import type { Permissions } from "../models";
 
+import type { DOOR_TOGGLE_MODE } from "./models";
+
 import { doorSystem } from ".";
 
 socket.on("Shape.Options.IsDoor.Set", (data: { shape: GlobalId; value: boolean }) => {
@@ -12,8 +14,14 @@ socket.on("Shape.Options.IsDoor.Set", (data: { shape: GlobalId; value: boolean }
     doorSystem.toggle(shape, data.value, SyncTo.UI);
 });
 
-socket.on("Shape.Options.DoorPermissions.Set", (data: { shape: GlobalId; value: Permissions }) => {
+socket.on("Shape.Options.Door.Permissions.Set", (data: { shape: GlobalId; value: Permissions }) => {
     const shape = getLocalId(data.shape);
     if (shape === undefined) return;
     doorSystem.setPermissions(shape, data.value, SyncTo.UI);
+});
+
+socket.on("Shape.Options.Door.ToggleMode.Set", (data: { shape: GlobalId; value: DOOR_TOGGLE_MODE }) => {
+    const shape = getLocalId(data.shape);
+    if (shape === undefined) return;
+    doorSystem.setToggleMode(shape, data.value, SyncTo.UI);
 });